requirements.txt
appdirs==1.4.4
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-simple-history==3.5.0
djangorestframework==3.15.1
frozendict==2.4.2
html5lib==1.1
idna==3.7
lxml==5.2.1
multitasking==0.0.11
numpy==1.26.4
pandas==2.2.2
peewee==3.17.3
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
setuptools==68.2.2
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
wheel==0.41.2
yfinance==0.2.38
